More room for protos, as expansion has not been implemented. Also saves storeheader on every aoaGC. --mg
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@9126 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/P/PStore.c
@@ -547,7 +547,7 @@ int setStoreObject(unsigned long storeID,
 				     delta * (currentPStore -> blockSize));
 	  
 	  Claim((long)blockOffset(offset)+ObjectSize(theObj)*4
-		< currentPStore->blockSize, 
+		< (long)currentPStore->blockSize, 
 		"Object is too large, it spans into the next block");
 
 	  memcpy(ObjectInStore,
